test(e2e): extract shared CLI harness and fixture builders
